внесены исправления по замечаниям
